aruivo pro X01/requirements.txt
httpx>=0.27
httpx[socks]
python-telegram-bot[socks]>=21.0
python-dotenv>=1.0
anthropic>=0.25
google-api-python-client>=2.120.0
